Entrega 3 + correciones

+++ Pet Lovers/src/presentation/users/services/eliminator-user.service.ts
@@ -1,6 +1,7 @@
 import { Repository } from "typeorm";
 import { User } from "../../../data/postgres/models/user.model";
 import { AppDataSource } from "../../../config/data-source";
+import { Response } from "express";
 
 export class DeleteUserService {
   private repository: Repository<User>;
@@ -9,19 +10,24 @@ export class DeleteUserService {
     this.repository = AppDataSource.getRepository(User);
   }
 
-  async deleteById(id: string) {
+  async deleteById(id: string, res: Response) {
+    if (!id || id.trim() === "" || id === ":id") {
+      return res.status(400).json({ message: "El ID es obligatorio" });
+    }
+
     try {
       const user = await this.repository.findOneBy({ id });
 
       if (!user) {
-        throw new Error("Usuario no encontrado");
+        return res.status(404).json({ message: "Usuario no encontrado" });
       }
 
       await this.repository.remove(user);
       console.log(`Usuario con ID ${id} eliminado correctamente.`);
+      return res.status(200).json({ message: "Usuario eliminado correctamente" });
     } catch (error) {
       console.error("Error al eliminar el usuario:", error);
-      throw new Error("No se pudo eliminar el usuario");
+      return res.status(500).json({ message: "No se pudo eliminar el usuario" });
     }
   }
 }
